refactor: harden static site daily refresh
